[refactor] create logo folder within assets and split up types for maintainability and scalability in the future
