+++ frontend/index.html
@@ -5,7 +5,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&family=DM+Sans:wght@400;500;600;700&display=swap" rel="stylesheet">
-    <link rel="icon" type="image/svg+xml" href="/frontend/src/assets/Exposr Logo.svg" />
+    <link rel="icon" type="image/svg+xml" href="/frontend/src/assets/logos/logo.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Exposr - AI Financial Risk Management Tool</title>
   </head>
