+++ frontend/src/App.tsx
@@ -1,4 +1,3 @@
-// Import styles of packages that you've installed.
 // All packages except `@mantine/hooks` require styles imports
 import '@mantine/core/styles.css';
 import { MantineProvider } from '@mantine/core';
