+++ frontend/src/types/position.ts
@@ -1,0 +1,7 @@
+export interface Position {
+  id: number;
+  portfolioId: number;
+  ticker: string;
+  quantity: number;
+  price: number;
+}
