+++ frontend/src/types/ui.ts
@@ -1,0 +1,7 @@
+import type { Portfolio } from "../types/portfolio";
+
+export interface PortfolioSelectProps {
+  portfolios: Portfolio[];
+  selectedId: number | null;
+  onSelect: (id: number) => void;
+}
